Adding Colour filtering program.
OpenCV is a library of programming functions mainly aimed at real-time computer vision.
This program will perform colour filtering according to the RGB you'll provide it. This program is actually going to convert your colors to HSV, which is "Hue Saturation Value."
You'll just have to provide the RGB value of the colour you want to filter from the input it's taking from your webcam.